requirements.txt
python-dotenv
pytz
requests
influxdb-client
pandas
google-generativeai
